# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mi primera documentacion con swagger",
        "description": "Si",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/pedidos": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Lista todos los pedidos.",
                "operationId": "18016b869e967852749747b26b6947ed",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_pedido": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "fecha_creacion": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "id_proveedor": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "id_forma_pago": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "id_cliente": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "id_estado_pedido": {
                                                        "type": "number",
                                                        "example": "1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Crea un nuevo pedido",
                "description": "Crea un nuevo pedido con los datos proporcionados en el cuerpo de la solicitud.",
                "operationId": "4e51d4584d4bd4f27c044e95086c820e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Fecha_creacion",
                                    "id_proveedor",
                                    "id_forma_pago",
                                    "id_cliente",
                                    "id_estado_pedido"
                                ],
                                "properties": {
                                    "Fecha_creacion": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-03-18"
                                    },
                                    "id_proveedor": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_forma_pago": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_cliente": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_estado_pedido": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pedido creado correctamente"
                                        },
                                        "pedido": {
                                            "$ref": "#/components/schemas/Pedidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud"
                    }
                }
            }
        },
        "/api/pedidos/{id}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Obtiene un pedido por su ID",
                "description": "Devuelve un pedido específico según el ID proporcionado.",
                "operationId": "664f26870b0021e2fe79a51482a3976d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Pedidos": {
                                            "$ref": "#/components/schemas/Pedidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pedido no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Actualiza un pedido.",
                "description": "Actualiza un pedido.",
                "operationId": "fab16f4160ac80507c7ddec0700cf878",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Fecha_creacion",
                                    "id_proveedor",
                                    "id_forma_pago",
                                    "id_cliente",
                                    "id_estado_pedido"
                                ],
                                "properties": {
                                    "Fecha_creacion": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-03-18"
                                    },
                                    "id_proveedor": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_forma_pago": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_cliente": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_estado_pedido": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pedido creado correctamente"
                                        },
                                        "pedido": {
                                            "$ref": "#/components/schemas/Pedidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Borra un pedido.",
                "description": "Borra un pedido.",
                "operationId": "2693c9e64eb791a7eaacf1cb48ae6cf8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Pedido eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pedido eliminado correctamente"
                                        },
                                        "pedido": {
                                            "$ref": "#/components/schemas/Pedidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Pedidos": {
                "title": "Pedidos",
                "description": "Modelo de datos para Pedidos.",
                "properties": {
                    "id_pedido": {
                        "type": "integer",
                        "example": "1"
                    },
                    "fecha_creacion": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-02-23T00:09:16.000000Z"
                    },
                    "id_proveedor": {
                        "type": "integer",
                        "example": "1"
                    },
                    "id_forma_pago": {
                        "type": "integer",
                        "example": "1"
                    },
                    "id_cliente": {
                        "type": "integer",
                        "example": "1"
                    },
                    "id_estado_pedido": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
        }
    }
}